Customizing Default Error Details
1. Build SampleProblemsFactory class, which implements ProblemDetailsFactory.

a. Implement CreateProblemDetails method, which instantiates a problem, adds values, adds the Title and Type if they are null, and returns the problem.
SampleProblemsFactory.cs: line 16-42

b. Implement CreateValidationProblemDetails, which returns a new ValidationProblemDetails
SampleProblemsFactory.cs: line 44-54

c. Add an ApiBehaviorOptions property and use it to provide access to ApiBehaviorOptions on the constructor.
SampleProblemsFactory.cs: line 10-15

2. Register the SampleProblemsFactory using AddTransient.
Program.cs: line 76